Algumas alterações, mas não implementei a paginação
